refactor(runtime): move worktree row resolution out of orca-runtime.ts
orca-runtime.ts is ~38k lines, so the scoped-resolution work should not add to
it. Move the row-resolution logic into `repo-worktree-row-resolution.ts`:
per-repo row building, the persisted-row fallback for a degraded scan, and the
scoped `id:` resolver.

The module owns no state. The two runtime-owned behaviours it needs -- the
cache-aware per-repo scan and folder-workspace identity stamping -- are injected
through `RepoWorktreeRowDeps`, so the scan cache stays where it lives today.

`withTimeout` moves to `src/shared/promise-timeout-fallback.ts` since both files
now need it.

Net effect: orca-runtime.ts is 113 lines shorter than on main, with the feature
added rather than removed.
