+++ src/main/runtime/orca-runtime.ts
@@ -734,6 +734,14 @@ import {
 } from './repo-worktree-resolution-scan'
 import { readRepoWorktreeAdminFingerprint } from './repo-worktree-admin-fingerprint'
 import {
+  listStoredWorktreeRowsForRepo,
+  resolveRepoWorktreeRows,
+  resolveScopedWorktreeIdRow,
+  RESOLVED_WORKTREE_REPO_TIMEOUT_MS,
+  type RepoWorktreeRowDeps
+} from './repo-worktree-row-resolution'
+import { withTimeout } from '../../shared/promise-timeout-fallback'
+import {
   getLocalWorktreePathAccess,
   removeLocalWorktreePath,
   toLocalWorktreeRuntimePath
@@ -29494,9 +29502,10 @@ export class OrcaRuntimeService {
         getAgentLaunchPlatformForRepo(repo, projectRuntimeByRepoId.get(repo.id))
       ])
     )
+    const deps = this.repoWorktreeRowDeps()
     const perRepoWorktrees = await Promise.all(
       repos.map(
-        async (repo) => await this.resolveRepoWorktreeRows(repo, metaById, projectRuntimeByRepoId)
+        async (repo) => await resolveRepoWorktreeRows(deps, repo, metaById, projectRuntimeByRepoId)
       )
     )
     const worktrees = projectResolvedWorktreeLineage(
@@ -29516,107 +29525,24 @@ export class OrcaRuntimeService {
     return { worktrees, platformByRepoId }
   }
 
-  /** One repo's worktree rows before lineage projection. Shared by the fleet scan and scoped lookups. */
-  private async resolveRepoWorktreeRows(
-    repo: Repo,
-    metaById: Record<string, ReturnType<NonNullable<Store['getWorktreeMeta']>>>,
-    projectRuntimeByRepoId: ReadonlyMap<string, ProjectExecutionRuntimeResolution>
-  ): Promise<ResolvedWorktree[]> {
-    if (isFolderRepo(repo)) {
-      return listRuntimeFolderWorkspaces(this.requireStore(), repo).map((worktree) => ({
-        ...worktree,
-        hostId: worktree.hostId ?? getRepoExecutionHostId(repo),
-        parentWorktreeId: null,
-        childWorktreeIds: [],
-        lineage: null,
-        git: {
-          path: worktree.path,
-          head: worktree.head,
-          branch: worktree.branch,
-          isBare: worktree.isBare,
-          isMainWorktree: worktree.isMainWorktree
-        },
-        displayName: worktree.displayName,
-        comment: worktree.comment
-      }))
+  /** Bind the runtime-owned scan cache and folder-workspace stamping into the row resolver. */
+  private repoWorktreeRowDeps(): RepoWorktreeRowDeps {
+    const store = this.requireStore()
+    return {
+      store,
+      scanRepo: (repo, projectRuntimeByRepoId) =>
+        this.listRepoWorktreesForResolution(repo, projectRuntimeByRepoId),
+      listFolderWorkspaces: (repo) => listRuntimeFolderWorkspaces(store, repo)
     }
-    // Why: mobile startup shares this path, so a slow repo scan degrades one repo's metadata instead of blocking all session loading.
-    // Why the catch: `withTimeout` resolves its fallback on rejection too, so the rejection must be absorbed first for `null` to mean
-    // "timed out" only. A stall never reached a verdict, so restore persisted rows instead of publishing a healthy-looking empty catalog;
-    // a rejection is a real answer and keeps its shipped zero-row semantics.
-    const scan: RuntimeWorktreeScanResult = (await withTimeout<RuntimeWorktreeScanResult | null>(
-      this.listRepoWorktreesForResolution(repo, projectRuntimeByRepoId).catch(
-        () => ({ ok: false, worktrees: [] }) satisfies RuntimeWorktreeScanResult
-      ),
-      RESOLVED_WORKTREE_REPO_TIMEOUT_MS,
-      null
-    )) ?? { ok: false, worktrees: this.listStoredWorktreesForResolution(repo) }
-    const gitWorktrees = scan.worktrees
-    if (scan.ok) {
-      pruneLineageForMissingRepoWorktrees(this.requireStore(), repo, gitWorktrees)
-    }
-    return gitWorktrees.map((gitWorktree) => {
-      const worktreeId = `${repo.id}::${gitWorktree.path}`
-      // Why: lineage validation needs a durable instance ID even when the runtime sees a workspace before renderer discovery-stamp.
-      const existingMeta = metaById[worktreeId]
-      const meta =
-        existingMeta && existingMeta.instanceId
-          ? existingMeta
-          : this.store?.setWorktreeMeta(worktreeId, {})
-      const merged = {
-        ...mergeWorktree(repo.id, gitWorktree, meta, repo.displayName),
-        hostId: existingMeta?.hostId ?? meta?.hostId ?? getRepoExecutionHostId(repo)
-      }
-      return {
-        ...merged,
-        parentWorktreeId: null,
-        childWorktreeIds: [],
-        lineage: null,
-        git: {
-          path: gitWorktree.path,
-          head: gitWorktree.head,
-          branch: gitWorktree.branch,
-          isBare: gitWorktree.isBare,
-          isMainWorktree: gitWorktree.isMainWorktree
-        },
-        displayName: merged.displayName,
-        comment: merged.comment
-      }
-    })
   }
 
-  /**
-   * Resolve one `<repoId>::<path>` worktree id by scanning only its owning repo.
-   *
-   * Lineage edges are intra-repo by construction (`sharesResolvedWorktreeLineageBoundary` requires a
-   * matching repoId), so projecting over one repo's rows yields the same parent and child ids the
-   * fleet scan would. Returns `null` whenever that does not hold, and the caller falls back.
-   */
   private async resolveExplicitWorktreeIdScoped(
     worktreeId: string
   ): Promise<ResolvedWorktree | null> {
-    const store = this.store
-    if (!store) {
+    if (!this.store) {
       return null
     }
-    const parsed = splitWorktreeIdForFilesystem(worktreeId)
-    if (!parsed?.repoId || !parsed.worktreePath) {
-      return null
-    }
-    const owners = store.getRepos().filter((repo) => repo.id === parsed.repoId)
-    // Why: one repo id can be registered on several execution hosts, and only the fleet scan
-    // decides between their rows. Hand those back to the unscoped path rather than guessing.
-    if (owners.length !== 1) {
-      return null
-    }
-    const repo = owners[0]
-    const rows = await this.resolveRepoWorktreeRows(
-      repo,
-      store.getAllWorktreeMeta() ?? {},
-      resolveLocalProjectRuntimesForRepos(this.requireStore(), [repo])
-    )
-    const projected = projectResolvedWorktreeLineage(rows, store.getAllWorktreeLineage?.() ?? {})
-    return projected.find((worktree) => worktree.id === worktreeId) ?? null
+    return await resolveScopedWorktreeIdRow(this.repoWorktreeRowDeps(), worktreeId)
   }
 
   private async listRepoWorktreesForResolution(
@@ -29776,37 +29702,7 @@ export class OrcaRuntimeService {
   }
 
   private listStoredWorktreesForResolution(repo: Repo): GitWorktreeInfo[] {
-    const store = this.store
-    if (!store) {
-      return []
-    }
-    const expectedHostId = getRepoExecutionHostId(repo)
-    const repoOwnerCount = store.getRepos().filter((candidate) => candidate.id === repo.id).length
-    const byWorktreeId = new Map<string, GitWorktreeInfo>()
-    for (const [worktreeId, meta] of Object.entries(store.getAllWorktreeMeta())) {
-      const parsed = splitWorktreeId(worktreeId)
-      if (!parsed || parsed.repoId !== repo.id) {
-        continue
-      }
-      // Why: one repo id can be registered on several execution hosts, so a degraded host must not republish another host's rows (same gate as worktrees.ts).
-      if (meta.hostId ? meta.hostId !== expectedHostId : repoOwnerCount > 1) {
-        continue
-      }
-      // Why: keep persisted rows for any repo kind while its scan is unreachable or stalled, instead of zero rows (worktrees:list does the same for disconnected SSH).
-      byWorktreeId.set(worktreeId, {
-        path: parsed.worktreePath,
-        head: '',
-        branch: '',
-        isBare: false,
-        isMainWorktree: areWorktreePathsEqual(parsed.worktreePath, repo.path),
-        ...(meta.sparseDirectories !== undefined ||
-        meta.sparseBaseRef !== undefined ||
-        meta.sparsePresetId !== undefined
-          ? { isSparse: true }
-          : {})
-      })
-    }
-    return [...byWorktreeId.values()]
+    return this.store ? listStoredWorktreeRowsForRepo(this.requireStore(), repo) : []
   }
 
   private async getResolvedWorktreeMap(): Promise<Map<string, ResolvedWorktree>> {
@@ -36137,7 +36033,6 @@ const WORKTREE_SCAN_AGENT_SCRATCH_TTL_MS = 5 * 60_000
 // edits are invisible to it and a tip living in packed-refs or reftable only gets an mtime + size
 // stamp, so a real scan still runs on this interval even while the probe reports "unchanged".
 export const WORKTREE_SCAN_ADMIN_RECONCILE_INTERVAL_MS = 5 * 60_000
-export const RESOLVED_WORKTREE_REPO_TIMEOUT_MS = 5000
 // Why reserved rather than spent on the probe: when the probe expires the caller still has to run
 // `git worktree list` and answer inside the same budget, so the fallback needs its own room. Sized
 // for a healthy Git on a busy host, well above the tens of milliseconds a warm list costs.
@@ -36234,21 +36129,6 @@ function getExplicitWorktreeIdSelector(selector: string | undefined): string | n
   }
   const id = selector.slice(3)
   return id.length > 0 ? id : null
-}
-
-function withTimeout<T>(promise: Promise<T>, timeoutMs: number, fallback: T): Promise<T> {
-  let timeout: ReturnType<typeof setTimeout> | null = null
-  return new Promise<T>((resolve) => {
-    timeout = setTimeout(() => resolve(fallback), timeoutMs)
-    promise.then(
-      (value) => resolve(value),
-      () => resolve(fallback)
-    )
-  }).finally(() => {
-    if (timeout) {
-      clearTimeout(timeout)
-    }
-  })
 }
 
 function withTimeoutResult<T>(
